Move restart methods into LifecycleUtils
