+++ app/src/org/commcare/activities/SingleAppManagerActivity.java
@@ -16,6 +16,7 @@ import org.commcare.logging.analytics.GoogleAnalyticsFields;
 import org.commcare.logging.analytics.GoogleAnalyticsUtils;
 import org.commcare.services.CommCareSessionService;
 import org.commcare.tasks.UpdateTask;
+import org.commcare.utils.LifecycleUtils;
 import org.commcare.utils.MultipleAppsUtil;
 import org.commcare.utils.SessionUnavailableException;
 import org.commcare.views.dialogs.StandardAlertDialog;
@@ -176,7 +177,7 @@ public class SingleAppManagerActivity extends CommCareActivity {
         GoogleAnalyticsUtils.reportAppManagerAction(GoogleAnalyticsFields.ACTION_UNINSTALL_APP);
         CommCareApplication.instance().expireUserSession();
         CommCareApplication.instance().uninstall(appRecord);
-        CommCareApplication.restartCommCare(
+        LifecycleUtils.restartCommCare(
                 SingleAppManagerActivity.this, AppManagerActivity.class, false);
     }
